fix: enhance error message formatting for localized Azure DevOps server errors
